Fixed the calculation of the function biWeeklyPay from int to double to make it compatible with polymorphism.

+++ Lab05/src/polymorphism/Millers.java
@@ -43,7 +43,7 @@ public String getName()
    @param hoursWorked the number of hours worked bi-weekly
    @return the pay for the given number of hours
 */
-public double biWeeklyPay(int hoursWorked)
+public double biWeeklyPay(double hoursWorked)
 {
    return 0; 
 }
